+++ src/commands/announce.js
@@ -22,14 +22,15 @@ const giftAction = async (interaction) => {
     .setTimestamp()
     .setTitle(personality.title)
     .setDescription(personality.description)
-    .setFooter(
-      {
-        text: personality.footer,
-        iconURL: "https://cdn.discordapp.com/avatars/691336942117453876/6d73900209e4d3bc35039f68f4aa9789.webp"
-      }
-    )
+    .setFooter({
+      text: personality.footer,
+      iconURL:
+        "https://cdn.discordapp.com/avatars/691336942117453876/6d73900209e4d3bc35039f68f4aa9789.webp",
+    })
     .addFields(Object.values(fields))
-    .setThumbnail("https://media.discordapp.net/attachments/959815577575256124/1041070360461852724/Ewilan_writing_cut.png?width=670&height=670");
+    .setThumbnail(
+      "https://media.discordapp.net/attachments/959815577575256124/1041070360461852724/Ewilan_writing_cut.png?width=670&height=670"
+    );
 
   //get channel
   const server = commons.find(({ guildId }) => guildId === interaction.guildId);
@@ -45,7 +46,7 @@ const giftAnnounce = {
   button: {
     name: "gift",
     value: "announce_gift",
-  }
+  },
 };
 
 // ANNOUNCE
@@ -54,8 +55,8 @@ const giftAnnounce = {
 const action = (interaction) => {
   // handle announce command interaction
 
-  const announceP = PERSONALITY.getCommands().announce;  //get personality
-  
+  const announceP = PERSONALITY.getCommands().announce; //get personality
+
   if (!isAdmin(interaction.user.id)) {
     //check for bot admin
     interactionReply(interaction, announceP.notAdmin);
@@ -85,14 +86,17 @@ const announces = [giftAnnounce]; //list of all announces
 //button action dispatcher
 export const announceButtonHandler = (interaction) => {
   const whichButton = interaction.customId;
-  const foundAnnounce = announces.find((obj) => obj.button.value === whichButton);
+  const foundAnnounce = announces.find(
+    (obj) => obj.button.value === whichButton
+  );
 
   if (foundAnnounce) foundAnnounce.action(interaction);
-  else interactionReply(interaction, PERSONALITY.getCommands().announce.notFound);
+  else
+    interactionReply(interaction, PERSONALITY.getCommands().announce.notFound);
 };
 
 //announce command
-const command = new SlashCommandBuilder() 
+const command = new SlashCommandBuilder()
   .setName(PERSONALITY.getCommands().announce.name)
   .setDescription(PERSONALITY.getCommands().announce.description)
   .setDefaultMemberPermissions(0x0000010000000000)
@@ -111,6 +115,9 @@ const announce = {
   help: (interaction) => {
     interactionReply(interaction, PERSONALITY.getCommands().announce.help);
   },
+  admin: true,
+  releaseDate: null,
+  sentinelle: true,
 };
 
 export default announce;