+++ src/commands/leadApo.js
@@ -81,6 +81,8 @@ const leaderboardApology = {
     return PERSONALITY.getCommands().leaderboardApology.help;
   },
   admin: true,
+  releaseDate: null,
+  sentinelle: true,
 };
 
 export default leaderboardApology;