+++ src/listeners.js
@@ -68,11 +68,15 @@ export const onPublicMessage = (message, client, currentServer, self) => {
       // spotify stuff
       checkIsOnThread(channel, playlistThreadId); //add bot if not on thread
     }
-    command.action(message, client, currentServer, self);
+    command.action(message, "$");
   }
 };
 
-export const onRemoveReminderReaction = (messageReaction, currentServer) => {
+export const onRemoveReminderReaction = (
+  messageReaction,
+  reactionUser,
+  currentServer
+) => {
   const { removeEmoji } = currentServer;
   const { message, emoji, users, client } = messageReaction;
 
@@ -80,12 +84,15 @@ export const onRemoveReminderReaction = (messageReaction, currentServer) => {
     // found corresponding reminder message
     ({ botMessage }) => botMessage.id === message.id
   );
+
   if (
     foundReminder &&
     emoji.name === removeEmoji &&
-    users.cache // if user reacting is the owner of reminder
-      .map((user) => user.id)
-      .includes(message.mentions.users.first().id)
+    (message.interaction
+      ? reactionUser.id === message.interaction.user.id
+      : users.cache
+          .map((user) => user.id)
+          .includes(message.mentions.users.first().id)) // if user reacting is the owner of reminder
   ) {
     try {
       client.remindme = client.remindme.filter(({ botMessage, timeout }) => {
@@ -93,6 +100,7 @@ export const onRemoveReminderReaction = (messageReaction, currentServer) => {
           // if it is the right message
           clearTimeout(timeout); //cancel timeout
           botMessage.reply(PERSONALITY.getCommands().reminder.delete);
+          console.log("reminder deleted");
           return false;
         }
         return true;
@@ -161,7 +169,7 @@ export const onReactionAdd = async (messageReaction, user) => {
 
   onRemoveSpotifyReaction(messageReaction, currentServer);
 
-  onRemoveReminderReaction(messageReaction, currentServer);
+  onRemoveReminderReaction(messageReaction, user, currentServer);
 };
 
 export const onReactionRemove = async (messageReaction, user) => {